Somehow missed the update earlier today
This should fix the issue people were having. a new param was added that affected how epochs were properly saved, but in Kohya fashion, they put it in the one place that isn't an easily callable function. In the future, the quickest way to reach me would be to use the github bug tracker.

Anyways, I added the param to the list of params and added it to the things you can mess with, but to be honest, it's better if you don't, the metadata it'll provide will be very helpful.