added json support, changing output file names, and reorganized
Did a bunch of things to the scripts that attempt to make it easier to find the params that you are looking for. basically weighted it from most important to least.
I also added the option that kohya added an update or two ago to allow for changing of outputs.
I also added support for saving and loading json files to make it easier for people to just kind of use other peoples params. Let me know if there is anything else that could be added, or report bugs to the bug tracker, I'm more likely to see it that way.