Big changes to the json loading and saving
Because I was getting complaints about the json system, I went ahead and completely changed it. now it's less readable to the average person, but that's fine, that isn't the goal now. To that end, I ended up changing pretty much everything else about the scripts.

so list of changes are:
* I made it so that everything is done in dicts, forgoing making an object of the ArgStore class and instead immediately grabbing a dict of it.
* Changed out how everything is loaded, for command_line, by default, it'll load everything. you must add to the new exclusion list that I have added, named json_load_skip_list. and for popup it by default excludes these following paramaters, base_model, img_folder, save_json_folder, reg_img_folder, lora_model_for_resume, change_output_name, training_comment, and the json_load_skip_list itself.
* Popup.py also got a few spelling fixes, and updated some of the defaults to better fit what the community is seemingly using.