updated everything to support the new loha implementation
this means making changes to the locon merge and extract scripts to support loha, as well as change variables to support it. LoCon as a repo was removed and replaced with LyCORIS. I also updated the popups to better reflect the options that are now available. since more than one thing uses the network_args and there are more args in general, I have deprecated the original locon implementation and instead started directly allowing the user to set the network_args instead.