updated to the newest version of sd-scripts and LyCORIS
With it, Updated the resize_lora.py to fit the new dynamic resizing, Updated the locon_extraction script to support the new sparce bais and cp decomposition.
I also added support for having two resolutions, with the main resolution still being named train_resolution handles either the width, or both if you don't define the new resolution variable height_resolution. I have added support for this in the popups as well as popups for the new cp decomposition for locon and loha.