Fixed bugs, added features
fixed a bug in which warmup ratio wasn't apply correctly. Added a button that allows you to add a root folder, and it will add all sub folders to the subset list. Changed the toml loading slightly so that it also sets the name of the subset based on the folder passed in. Fixed the scrolling issue with tabs