lots of changes
Changed the way the sections are displayed, changing them into tabs to allow for much smaller screen size. Save the last folder loaded when you load or save a toml file. reorganized the save args to make it mor compact. added a folder selection to the save toml and tag occurence files. set it up to automatically change the file names if it already exists, rather than overwrite it.