changed position of queue to better reflect screen size.
further reorganized the save args.
removed lyco dropout, as it was just network dropout.
added lycoris dylora.
updated lycoris to support the new dropout conditions.
refactored the algo select to be more maintainable in the future.
changed the behavior of enable and disable dropouts to better relate to how they work now.